+++ extensions/kill-process/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Kill Process Changelog
 
+## [Sort by memory and App aggregation] - 2023-05-27
+
+- Now can sort processes by CPU or memory usage
+- Now can aggregate processes of the same app
+
 ## [Filter by PID] - 2023-03-20
 
 - Add filtering pid functionality